HandleSpawnItems seperate class
