+++ Willi.PlateUpEnhancementMod/Config/ConfigHelper.cs
@@ -2,6 +2,7 @@
 using BepInEx.Logging;
 using System.Collections.Generic;
 using UnityEngine;
+using Willi.PlateUpEnhancementMod.Helpers;
 
 namespace Willi.PlateUpEnhancementMod.Config
 {
@@ -314,7 +315,7 @@ namespace Willi.PlateUpEnhancementMod.Config
         private static ConfigFile BindItemSpawnerConfig(this ConfigFile config)
         {
             SpawnItemId = config.Bind("1. Spawn Any Item", "ItemIdToSpawn:", "-571205127",
-                new ConfigDescription("Enter an item id to spawn it", null, new ConfigurationManagerAttributes { CustomDrawer = PlateUpEnhancement.HandleSpawnItemConfigManager, Order = 100 })
+                new ConfigDescription("Enter an item id to spawn it", null, new ConfigurationManagerAttributes { CustomDrawer = HandleSpawnItems.HandleSpawnItemConfigManager, Order = 100 })
             );
             SpawnItemPrice = config.Bind("1. Spawn Any Item", "PriceOfItem", 10, new ConfigDescription("The price of the item when spawned", null, new ConfigurationManagerAttributes { Order = 90 }));
             SpawnItemKeyboardShortcut = config.Bind("1. Spawn Any Item", "KeyboardShortcut", new KeyboardShortcut(KeyCode.J), new ConfigDescription("Use this keyboard shortcut to spawn the item", null, new ConfigurationManagerAttributes { Order = 80 }));
